+++ R/calibrate.R
@@ -168,32 +168,20 @@ calib.plot <- function(set=get('info'), BCAD=set$BCAD, cc=set$cc, rotate.axes=FA
       yrsteps <- min(yrsteps, diff(set$calib$probs[[1]][,1]))
       }
 
-yrsteps <<- yrsteps      
-      
   if(plot.dists)
     for(i in 1:length(set$calib$probs)) {
       cal <- cbind(set$calib$probs[[i]])
       d <- set$calib$d[[i]]
       if(BCAD)
         cal[,1] <- 1950-cal[,1]
-      o <- order(cal[,1])
-      cal <- cbind(cal[o,1], cal[o,2])
       cal <- approx(cal[,1], cal[,2], seq(min(cal[,1]), max(cal[,1]), by=yrsteps)) # all dists should have the same binsize
-      cal <- cbind(cal$x, cal$y)
+      cal <- cbind(cal$x, cal$y/sum(cal$y))
       if(same.heights)
-        cal[,2] <- cal[,2]/max(cal[,2]) else
-          cal[,2] <- cal[,2]/maxhght
-      if(normalise.dists)
-        cal[,2] <- cal[,2]/sum(cal[,2])
-      #cal[,2] <- height*cal[,2] # new definition below, 31 Jan 2021
-	  cal[,2] <- cal[,2]*height*(max(dlim) - min(dlim))/(5*maxhght) # the value 5 should result in visible blobs in most cases. Otherwise adapt with the option height
-      if(ncol(set$dets) > 4 && set$dets[i,5] == 0) # cal BP date
-        cal[,2] <- calheight*cal[,2]
- 
- 
- 
- # cal BP dates need same amount of total area as C14 dates... How?
-  
+        cal[,2] <- cal[,2]/max(cal[,2])/3 # /3 because next line we do *3
+      cal[,2] <- 3 * height * cal[,2] * (max(dlim) - min(dlim)) # scale the height of the blobs with the core length
+      if(ncol(set$dets) > 4 && set$dets[i,5] == 0) # cal BP dates could have different relative height:
+        cal[,2] <- calheight * cal[,2]
+
       if(mirror)
         pol <- cbind(c(d-cal[,2], d+rev(cal[,2])), c(cal[,1], rev(cal[,1]))) else
          if(up)
@@ -255,14 +243,12 @@ yrsteps <<- yrsteps
     if(set$normal)
       cal <- cbind(cc[,1], dnorm(cc[,2], rcmean, sqrt(cc[,3]^2+w2))) else
         cal <- cbind(cc[,1], (t.b+ ((rcmean-cc[,2])^2) / (2*(cc[,3]^2 + w2))) ^ (-1*(t.a+0.5))) # student-t
-    cal[,2] <- cal[,2]/max(cal[,2]) # all peaks at 1
-    these <- which(cal[,2] >= cutoff) # keep values >% of the maximum height  
-    cal <- cal[these,]
-    #cal[,2] <- cal[,2]/sum(cal[,2]) # now normalise	
-
-    calx <- seq(min(cal[,1]), max(cal[,1]), length=date.res)
-    caly <- approx(cal[,1], cal[,2], calx)$y
-    cbind(calx, caly/sum(caly), deparse.level = 0)
+    cal[,2] <- cal[,2] / sum(cal[,2]) # normalise
+    
+    cal.left <- min(which(cal[,2]/max(cal[,2]) >= cutoff)) # remove outer bits
+    cal.right <- nrow(cal) - min(which(cal[nrow(cal):1,2]/max(cal[,2]) >= cutoff)) # ... here too
+     #   cat(i, ", ", cal[cal.left,1], ", ", cal[cal.right,1], "\n")
+    return(cal[cal.left:cal.right,])
   }
 
   # now calibrate all dates
@@ -283,7 +269,7 @@ yrsteps <<- yrsteps
       for(i in 1:nrow(dat)) {
         dets <- c(NA, as.numeric(dat[i,-1])) # the first column is not numeric
         if(dets[5]==0) {
-          x <- seq(dets[2]-(5*dets[3]), dets[2]+(5*dets[3]), length=date.res)
+          x <- seq(dets[2]-(4*dets[3]), dets[2]+(4*dets[3]), length=date.res)
           ccurve <- cbind(x, x, rep(0,length(x))) # dummy 1:1 curve
         } else {
             if(dets[5]==1) ccurve <- cc1 else if(dets[5]==2) ccurve <- cc2 else