+++ NEWS.md
@@ -1,7 +1,8 @@
 # rbacon 2.5.1
 * pMC.age and other IntCal functions should now load as expected
-* Adapted the default prior for memory to 0.5 (mean) and 10 (strength), to repair a bug with the original bacon c++ code. This default should work with most cores and give similar results to the previous settings for the memory prior.
+* Adapted the default prior for memory to 0.5 (mean) and 10 (strength), to repair a bug with the original bacon c++ code. This default should work with most cores and give similar results to the previous settings for the memory prior
 * Updated c/c++ code (as used in version 2.4.1 with some minor additional updates) 
+* the heights of the calibrated distributions should now scale better according to how precise they are
 
 
 # rbacon 2.5.0